feat: added polls apps models
Added  models for Question and Choice to the models app, The question model holds the raised
issue for a poll, while the choices represent a single option beloning to a question.
Question model has also been enabled for editing within the admin section of django.
This commit also contains the commits made